requirements.txt
altair==5.0.1
anaconda-client==1.11.2
anaconda-navigator==2.4.0
anaconda-project==0.11.1
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
AudioRecorder==0.1
Babel==2.11.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.6.2
boltons==23.0.0
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clyent==1.2.2
colorama==0.4.6
comm==0.2.1
conda==23.3.1
conda-build==3.24.0
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
conda-repo-cli==1.0.88
conda-token==0.4.0
conda-verify==3.4.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
fonttools==4.53.1
future==0.18.3
gitdb==4.0.7
GitPython==3.1.43
glob2==0.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.28.0
ipython==8.25.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==1.0.0
kiwisolver==1.4.5
libarchive-c==2.9
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.0
menuinst==1.4.19
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
navigator-updater==0.5.1
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.8
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
openai==1.37.1
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pillow==10.4.0
pip==22.3.1
pkginfo==1.10.0
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==16.1.0
PyAudio==0.2.14
pycosat==0.6.6
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.0
pydub==0.25.1
pyglet==2.0.16
Pygments==2.15.1
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.0.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
QtPy==2.4.1
referencing==0.30.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
Send2Trash==1.8.2
setuptools==65.6.3
sip==6.7.12
six==1.16.0
smmap==4.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
streamlit==1.32.0
streamlit-audiorecorder==0.0.5
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.38.4
widgetsnbextension==3.6.6
win-inet-pton==1.1.0
wincertstore==0.2
zstandard==0.22.0
